Do not populate repeater fields in admin with SEO form
If a Static page layout extends form fields of a page with a repeater field then fields included in the repeater are also populated with the SEO fields from this plugin. The code needs to detect the level of form fields and stop populating its forms if deeper than first level.

Fix for Issue #2 in original repo.