Make Translations and multilevel page structure work
The original code relies on a flat page structure. If a website uses deeper page structure then simply using content after first slash will not work. This code strips first three letters from the $url (eg. "/pl") and leaves the remainder untouched for page detection. Also detects if we're at the website root and if so then adds "/" back to the $url.

To get translated data from viewBag property it is necessary to use $page->viewBag() instead of $page->getViewBag() which is (according to docs) for internal use only. Remaining changes to $viewBag['xy'] are necessary because the $page->viewBag() returns an array instead of an object.